♻️ Refactor: Dashboard 내 STEP과 TASK 타입 분리
