+++ src/types/api/tasks.ts
@@ -1,7 +1,36 @@
-// TaskItem 컴포넌트 관련 타입 정의
+import { ApiErrorResponse } from './error';
 
-// 기본 Task 인터페이스
+// API용 Task 타입 (완전한 형태)
 export interface Task {
+  taskId: number;
+  taskName: string;
+  status: 'ONGOING' | 'COMPLETED' | 'PENDING';
+  managers: Manager[];
+  deadline: string;
+}
+
+// 담당자 타입
+export interface Manager {
+  userId: number;
+  userName: string;
+}
+
+// TASK 생성 요청 타입
+export interface CreateTaskRequest {
+  stepId: number;
+}
+
+// TASK 생성 응답 타입
+export interface CreateTaskResponse {
+  isSuccess: boolean;
+  error: ApiErrorResponse | null;
+  result: {
+    taskId: number;
+  } | null;
+}
+
+// UI용 Task 타입 (간단한 형태)
+export interface TaskItem {
   id: number;
   title: string;
   status: string;
@@ -10,13 +39,7 @@ export interface Task {
 }
 
 // TaskItem 컴포넌트 Props
-export interface TaskItemProps {
-  id: number;
-  title: string;
-  status: string; // 유연성을 위해 string으로 변경
-  deadline?: string;
-  assignee?: string[];
-}
+export interface TaskItemProps extends TaskItem {}
 
 // TaskItem 컴포넌트 Props (projectId 포함)
 export interface TaskItemComponentProps extends TaskItemProps {
@@ -25,7 +48,7 @@ export interface TaskItemComponentProps extends TaskItemProps {
 
 // useTaskItems 훅 Props
 export interface UseTaskItemsProps {
-  task: Task;
+  task: TaskItem;
 }
 
 // useTaskItems 훅 반환값
@@ -40,17 +63,17 @@ export interface UseTaskItemsReturn {
   deadlineTextColor: string;
 }
 
-// Task 상태 매핑 타입
+// TASK 상태 매핑 타입
 export type TaskStatus = 'BEFORE' | 'ONGOING' | 'COMPLETE';
 
-// Task 상태 표시 텍스트 매핑
+// TASK 상태 표시 텍스트 매핑
 export const TASK_STATUS_DISPLAY: Record<TaskStatus, string> = {
   BEFORE: '시작 전',
   ONGOING: '진행 중',
   COMPLETE: '완료',
 } as const;
 
-// Task 상태별 스타일 매핑
+// TASK 상태별 스타일 매핑
 export const TASK_STATUS_STYLES: Record<string, { bg: string; text: string }> = {
   '진행 중': { bg: 'bg-[#B6F5DF]', text: 'text-[#505050]' },
   완료: { bg: 'bg-[#D1D5DB]', text: 'text-[#505050]' },
